[exporter/otlp] split into its own module
This is to address https://github.com/open-telemetry/opentelemetry-collector/issues/6195
