+++ cmd/builder/internal/config/default.yaml
@@ -11,8 +11,7 @@ receivers:
 exporters:
   - import: go.opentelemetry.io/collector/exporter/loggingexporter
     gomod: go.opentelemetry.io/collector v0.62.1
-  - import: go.opentelemetry.io/collector/exporter/otlpexporter
-    gomod: go.opentelemetry.io/collector v0.62.1
+  - gomod: go.opentelemetry.io/collector/exporters/otlpgrpcexporter v0.62.1
   - import: go.opentelemetry.io/collector/exporter/otlphttpexporter
     gomod: go.opentelemetry.io/collector v0.62.1
 extensions:
