+++ cmd/otelcorecol/go.mod
@@ -7,6 +7,7 @@ go 1.18
 require (
 	github.com/stretchr/testify v1.8.0
 	go.opentelemetry.io/collector v0.62.1
+	go.opentelemetry.io/collector/exporters/otlpgrpcexporter v0.62.1
 	golang.org/x/sys v0.0.0-20220919091848-fb04ddd9f9c8
 )
 
@@ -69,9 +70,9 @@ require (
 	go.uber.org/atomic v1.10.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
 	go.uber.org/zap v1.23.0 // indirect
-	golang.org/x/net v0.0.0-20220225172249-27dd8689420f // indirect
+	golang.org/x/net v0.0.0-20220909164309-bea034e7d591 // indirect
 	golang.org/x/text v0.3.7 // indirect
-	google.golang.org/genproto v0.0.0-20211208223120-3a66f561d7aa // indirect
+	google.golang.org/genproto v0.0.0-20220930163606-c98284e70a91 // indirect
 	google.golang.org/grpc v1.50.1 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
@@ -80,6 +81,8 @@ require (
 
 replace go.opentelemetry.io/collector => ../../
 
-replace go.opentelemetry.io/collector/semconv => ../../semconv
+replace go.opentelemetry.io/collector/exporters/otlpgrpcexporter => ../../exporters/otlpgrpcexporter
 
 replace go.opentelemetry.io/collector/pdata => ../../pdata
+
+replace go.opentelemetry.io/collector/semconv => ../../semconv
