+++ exporters/otlpgrpcexporter/doc.go
@@ -13,4 +13,4 @@
 // limitations under the License.
 
 // Package otlpexporter exports data by using the OTLP format to a gPRC endpoint.
-package otlpexporter // import "go.opentelemetry.io/collector/exporter/otlpexporter"
+package otlpgrpcexporter // import "go.opentelemetry.io/collector/exporters/otlpgrpcexporter"
